Move processes to rails_application because they are application-specific and we want them defined in the app they belong to.
