--- ecommerce/processes/.mutant.yml
@@ -1,20 +1,0 @@
-requires:
-  - ./test/test_helper
-integration: minitest
-usage: opensource
-coverage_criteria:
-  process_abort: true
-matcher:
-  subjects:
-    - Processes*
-  ignore:
-    - Processes::Configuration*
-    - Processes::OrderConfirmation#stream_name
-    - Processes::Test*
-    - Processes::ReleasePaymentProcess*
-    - Processes::MoneySplitter*
-    - Processes::OrderItemInvoicingProcess#fetch_id
-    - Processes::SyncShipmentFromPricing*
-    - Processes::SyncInventoryFromOrdering*
-    - Processes::NotifyPaymentsAboutOrderValue*
-    - Processes::WelcomeMessageProcess*
