--- ecommerce/processes/Gemfile
@@ -1,4 +1,0 @@
-source "https://rubygems.org"
-
-eval_gemfile "../../infra/Gemfile.test"
-gem "infra", path: "../../infra"
