+++ rails_application/app/processes/processes/configuration.rb
@@ -1,24 +1,3 @@
-require_relative "../../ordering/lib/ordering"
-require_relative "../../pricing/lib/pricing"
-require_relative "../../product_catalog/lib/product_catalog"
-require_relative "../../crm/lib/crm"
-require_relative "../../payments/lib/payments"
-require_relative "../../inventory/lib/inventory"
-require_relative "../../shipping/lib/shipping"
-require_relative "../../taxes/lib/taxes"
-require_relative "../../invoicing/lib/invoicing"
-require_relative "../../fulfillment/lib/fulfillment"
-require_relative 'processes/confirm_order_on_payment_captured'
-require_relative 'processes/release_payment_process'
-require_relative 'processes/shipment_process'
-require_relative 'processes/determine_vat_rates_on_order_placed'
-require_relative 'processes/order_item_invoicing_process'
-require_relative 'processes/notify_payments_about_order_value'
-require_relative 'processes/sync_shipment_from_pricing'
-require_relative 'processes/three_plus_one_free'
-require_relative 'processes/reservation_process'
-require_relative 'processes/welcome_message_process'
-
 module Processes
   class Configuration
     class << self
