+++ rails_application/app/processes/processes/notify_payments_about_order_value.rb
