+++ rails_application/app/processes/processes/welcome_message_process.rb
