+++ rails_application/test/processes/money_splitter_test.rb
@@ -1,4 +1,4 @@
-require_relative "test_helper"
+require "test_helper"
 
 module Processes
   class MoneySplitterTest < Minitest::Test
